Rename TypeScript files to kebab-case and put all tests under tests/unit/
