+++ src/rules/index.ts
@@ -1,9 +1,9 @@
 import type {Rule} from '../types.js'
 import {filenameAltText} from './filename-alt-text.js'
-import {vagueAltText} from './vagueAltText.js'
-import {missingAltText} from './missingAltText.js'
+import {vagueAltText} from './vague-alt-text.js'
+import {missingAltText} from './missing-alt-text.js'
 import {placeholderAltText} from './placeholder-alt-text.js'
-import {repeatedAltText} from './repeatedAltText.js'
+import {repeatedAltText} from './repeated-alt-text.js'
 
 // Append-only registry. Add a rule by importing it here and pushing it onto the array.
 export const allRules: Rule[] = [filenameAltText, vagueAltText, missingAltText, placeholderAltText, repeatedAltText]
