+++ src/rules/missing-alt-text.ts
