+++ src/utils/normalize-alt-text.ts
